Code improvements in ctxt step

+++ IM/ConfManager.py
@@ -95,7 +95,7 @@ class ConfManager(threading.Thread):
 	def stop(self):
 		self._stop = True
 		# put a task to assure to wake up the thread 
-		self.inf.ctxt_tasks.put((-3, 0,None,None))
+		self.inf.ctxt_tasks.put((-10, 0,None,None))
 		ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Stop Configuration thread.")
 
 	def check_vm_ips(self, timeout = Config.WAIT_RUNNING_VM_TIMEOUT):
@@ -138,6 +138,14 @@ class ConfManager(threading.Thread):
 				
 		return success
 
+	def kill_ctxt_processes(self):
+		"""
+			Kill all the ctxt processes 
+		"""
+		for vm in self.inf.get_vm_list():
+			ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Killing ctxt processes.")
+			vm.kill_check_ctxt_process()
+
 	def run(self):
 		ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Starting the ConfManager Thread")
 
@@ -148,8 +156,7 @@ class ConfManager(threading.Thread):
 			if self.init_time + self.max_ctxt_time < time.time():
 				ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Max contextualization time passed. Exit thread.")
 				# Kill the ansible processes
-				for vm in self.inf.get_vm_list():
-					vm.kill_check_ctxt_process()
+				self.kill_ctxt_processes()
 				return
 
 			vms_configuring = self.check_running_pids(vms_configuring)
@@ -237,34 +244,41 @@ class ConfManager(threading.Thread):
 			ip = vm.getPublicIP()
 			if not ip:
 				ip = vm.getPrivateIP()
-			remote_dir = Config.REMOTE_CONF_DIR + "/" + ip + "_" + str(vm.getSSHPort())
-			tmp_dir = tempfile.mkdtemp()
-	
-			ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Create the configuration file for the contextualization agent")
-			conf_file = tmp_dir + "/config.cfg"
-			self.create_vm_conf_file(conf_file, vm.im_id, tasks, remote_dir)
-			
-			ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Copy the contextualization agent config file")
-	
-			# Copy the contextualization agent config file
-			ssh = self.inf.vm_master.get_ssh(retry = True)
-			ssh.sftp_mkdir(remote_dir)
-			ssh.sftp_put(conf_file, remote_dir + "/" + os.path.basename(conf_file))
-			
-			shutil.rmtree(tmp_dir, ignore_errors=True)
-	
-			(pid, _, _) = ssh.execute("nohup python_ansible " + Config.REMOTE_CONF_DIR + "/ctxt_agent.py " 
-					+ Config.REMOTE_CONF_DIR + "/general_info.cfg "
-					+ remote_dir + "/" + os.path.basename(conf_file) 
-					+ " > " + remote_dir + "/stdout" + " 2> " + remote_dir + "/stderr < /dev/null & echo -n $!")
-			
-			ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Ansible process to configure " + str(vm.im_id) + " launched with pid: " + pid)
-			
-			vm.ctxt_pid = pid
-			vm.launch_check_ctxt_process()
+
+			if not ip:
+				ConfManager.logger.error("Inf ID: " + str(self.inf.id) + ": VM with ID %s (%s) does not have an IP!!. We cannot launch the ansible process!!" % (str(vm.im_id), vm.id))
+			else:
+				remote_dir = Config.REMOTE_CONF_DIR + "/" + ip + "_" + str(vm.getSSHPort())
+				tmp_dir = tempfile.mkdtemp()
+		
+				ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Create the configuration file for the contextualization agent")
+				conf_file = tmp_dir + "/config.cfg"
+				self.create_vm_conf_file(conf_file, vm.im_id, tasks, remote_dir)
+				
+				ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Copy the contextualization agent config file")
+		
+				# Copy the contextualization agent config file
+				ssh = self.inf.vm_master.get_ssh(retry = True)
+				ssh.sftp_mkdir(remote_dir)
+				ssh.sftp_put(conf_file, remote_dir + "/" + os.path.basename(conf_file))
+				
+				shutil.rmtree(tmp_dir, ignore_errors=True)
+		
+				if vm.configured is None:
+					(pid, _, _) = ssh.execute("nohup python_ansible " + Config.REMOTE_CONF_DIR + "/ctxt_agent.py " 
+							+ Config.REMOTE_CONF_DIR + "/general_info.cfg "
+							+ remote_dir + "/" + os.path.basename(conf_file) 
+							+ " > " + remote_dir + "/stdout" + " 2> " + remote_dir + "/stderr < /dev/null & echo -n $!")
+				
+					ConfManager.logger.debug("Inf ID: " + str(self.inf.id) + ": Ansible process to configure " + str(vm.im_id) + " launched with pid: " + pid)
+					
+					vm.ctxt_pid = pid
+					vm.launch_check_ctxt_process()
+				else:
+					ConfManager.logger.warn("Inf ID: " + str(self.inf.id) + ": Ansible process to configure " + str(vm.im_id) + " NOT launched")
 		except:
 			pid = None
-			ConfManager.logger.exception("Inf ID: " + str(self.inf.id) + ": Error launching the ansible process to configure %s" % str(vm.im_id))
+			ConfManager.logger.exception("Inf ID: " + str(self.inf.id) + ": Error launching the ansible process to configure VM with ID %s" % str(vm.im_id))
 
 		# If the process is not correctly launched the configuration of this VM fails
 		if pid is None:
